Refactor publisher (#158)
